Fixed bugs when saving a profile photo - new and existing users.

+++ Models/PhotoManager.cs
@@ -47,15 +47,19 @@ namespace BodyCompositionCalculator.Models
             _context.SaveChanges();
             _context.UserProfiles.SingleOrDefault(m => m.Id == userProfileId).UserPhotoId = newPhoto.Id;
             _context.SaveChanges();
-            DeletePhotoFromDb(userProfilePhoto.Id);
+ 
 
         }
 
         public void DeletePhotoFromDb(int id)
         {
             var photo = _context.UserPhotos.SingleOrDefault(m => m.Id == id);
-            _context.UserPhotos.Remove(photo);
-            _context.SaveChanges();
+            if (photo != null)
+            {
+                _context.UserPhotos.Remove(photo);
+                _context.SaveChanges();
+            }
+            
         }
     }
 }
